GH-1245 Move buffer_factory into buffer_factor.hpp
